requirements.txt
Flask==3.1.0
tensorflow==2.19.0
gunicorn==20.1.0
Pillow==9.5.0
numpy==1.24.3